refactor: result api, react-query data훅으로 수정 및 적용
