+++ apps/frontend/src/_components/pages/interview/RecordButton/RecordButton.tsx
@@ -24,11 +24,7 @@ import { createRecordError, RecordErrorType } from "./_error";
 import useAlertDialogStore from "@repo/store/useAlertDialogStore";
 import useConfirmDialogStore from "@repo/store/useConfirmDialogStore";
 import useToastStore from "@repo/store/useToastStore";
-import { useMutation } from "@tanstack/react-query";
-import { delay } from "@/_libs/utils";
-import { APIError } from "@/_apis/fetcher";
-import { fetchCreateResult } from "@/_apis/result";
-import { useParams, useRouter } from "next/navigation";
+import { useCreateResult } from "@/_data/result";
 
 export enum RecordingStatusType {
   idle = "idle",
@@ -37,8 +33,6 @@ export enum RecordingStatusType {
 }
 
 export default function RecordButton() {
-  const router = useRouter();
-  const interviewId = useParams().interviewId;
   const recorderRef = useRef<Recorder | null>(null);
   const [recordingStatus, setRecordingStatus] = useState<RecordingStatusType>(
     RecordingStatusType.idle
@@ -56,40 +50,7 @@ export default function RecordButton() {
   const { setAlert } = useAlertDialogStore();
   const { setConfirm } = useConfirmDialogStore();
   const { addToast } = useToastStore();
-  const { mutate, isPending } = useMutation({
-    mutationFn: async () => {
-      await delay(500);
-
-      return await fetchCreateResult({
-        interviewId: Number(interviewId),
-      });
-    },
-    onSuccess: (data) => {
-      if (!data?.id) {
-        addToast({
-          title: "인터뷰 결과 생성 실패",
-          description: "인터뷰 결과 생성에 실패했습니다. 다시 시도해주세요.",
-        });
-        return;
-      }
-
-      router.push(`/result/${data.id}`);
-    },
-    onError: (error) => {
-      if (error instanceof APIError) {
-        addToast({
-          title: "인터뷰 결과 생성 실패",
-          description: error.message,
-        });
-        return;
-      }
-
-      addToast({
-        title: "인터뷰 결과 생성 실패",
-        description: "인터뷰 결과 생성에 실패했습니다. 다시 시도해주세요.",
-      });
-    },
-  });
+  const { mutate, isPending } = useCreateResult();
 
   const handleRecord = async () => {
     try {
