+++ apps/frontend/src/_components/pages/result/TotalEvaluationSection.tsx
@@ -1,18 +1,11 @@
 import styles from "./TotalEvaluationSection.module.css";
 import Text from "@repo/ui/Text";
-import { useParams } from "next/navigation";
-import { useQuery } from "@tanstack/react-query";
-import { fetchGetResultTotalEvaluation } from "@/_apis/result";
+import { useGetResultTotalEvaluation } from "@/_data/result";
 
 interface TotalEvaluationSectionProps {}
 
 export default function TotalEvaluationSection({}: TotalEvaluationSectionProps): React.ReactElement {
-  const { resultId } = useParams();
-  const { data, isLoading, error } = useQuery({
-    queryKey: ["result", resultId, "total"],
-    queryFn: () =>
-      fetchGetResultTotalEvaluation({ resultId: Number(resultId) }),
-  });
+  const { data, isLoading, error } = useGetResultTotalEvaluation();
 
   if (isLoading) {
     return <div>Loading...</div>;
